requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
boto3>=1.26.0  # AWS SDK for Python
python-dotenv>=0.19.0  # For managing environment variables
requests>=2.28.0
psycopg2-binary>=2.9.3  # PostgreSQL adapter for Python
redshift-connector>=2.0.0  # Add this line 
pymongo==3.12.3 